refactor: 프로젝트 구조 변경 및 일부 리팩터링 필요한 코드 삭제
Schedule package OpenAPI 와 분리
BusLocation 이름 BusRouteLocation 으로 변경
UserRequest name 컬럼 삭제
