+++ bus-seat/src/main/java/org/kernel360/busseat/route_location/controller/BusLocationController.java
@@ -1,0 +1,13 @@
+package org.kernel360.busseat.route_location.controller;
+
+import org.springframework.web.bind.annotation.RestController;
+
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+
+@Slf4j
+@RestController
+@RequiredArgsConstructor
+public class BusLocationController {
+	// have to implement this
+}
