+++ bus-seat/src/main/java/org/kernel360/busseat/user_request/dto/UserRequestDto.java
@@ -12,8 +12,5 @@ import lombok.Setter;
 @NoArgsConstructor
 @AllArgsConstructor
 public class UserRequestDto {
-
 	private Long routeId;
-
-	private String routeName;
 }
